Добавил README.md, поменял подключение к бд
