+++ src/main/java/ru/teamscore/java23/CurrencyApplication.java
@@ -1,11 +1,48 @@
 package ru.teamscore.java23;
 
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+import org.springframework.web.context.annotation.RequestScope;
+
+import javax.sql.DataSource;
+
 
 @SpringBootApplication
 public class CurrencyApplication {
     public static void main(String[] args) {
         SpringApplication.run(CurrencyApplication.class, args);
+    }
+
+    @Bean
+    public DataSource dataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName("org.postgresql.Driver");
+        dataSource.setUrl("jdbc:postgresql://localhost:5432/currencies");
+        dataSource.setUsername("postgres");
+        dataSource.setPassword("1234");
+        return dataSource;
+    }
+
+    //Проблемы с видимостью классов при использовании hibernate.cfg.xml
+    //Сделал напрямую и указал сканирование нужного модуля
+    @Bean
+    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
+        LocalContainerEntityManagerFactoryBean em = new LocalContainerEntityManagerFactoryBean();
+        em.setDataSource(dataSource());
+        em.setPackagesToScan("ru.teamscore.java23.models");
+        em.setJpaVendorAdapter(new HibernateJpaVendorAdapter());
+        return em;
+    }
+
+    @Bean
+    @RequestScope // Установка области видимости запроса
+    public EntityManager entityManager(EntityManagerFactory entityManagerFactory) {
+        return entityManagerFactory.createEntityManager();
     }
 }
