--- src/main/resources/hibernate.cfg.xml
@@ -1,12 +1,0 @@
-<hibernate-configuration xmlns="http://www.hibernate.org/xsd/orm/cfg">
-    <session-factory>
-        <property name="hibernate.connection.driver_class">org.postgresql.Driver</property>
-        <property name="hibernate.connection.url">jdbc:postgresql://localhost/currencies</property>
-        <property name="hibernate.connection.username">currencies_user</property>
-        <property name="hibernate.connection.password">currencies</property>
-        <!-- Print executed SQL to stdout -->
-        <property name="show_sql">true</property>
-        <!-- Create/update database structure on start -->
-        <property name="hibernate.hbm2ddl.auto">validate</property>
-    </session-factory>
-</hibernate-configuration>
