+++ src/main/resources/templates/admin.html
@@ -52,15 +52,15 @@
         <div class="collapse navbar-collapse" id="navbarNav">
           <ul class="navbar-nav">
             <li class="nav-item">
-              <a class="nav-link active" aria-current="page" href="admin.html"
+              <a class="nav-link active" aria-current="page" href="admin/currency/index"
                 >Валюты</a
               >
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="admin/currency_pair/index.html">Валютные пары</a>
+              <a class="nav-link" href="admin/currency_pair/index">Валютные пары</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="admin/exchange_rate/index.html">Обменный курс</a>
+              <a class="nav-link" href="admin/exchange_rate/index">Обменный курс</a>
             </li>
           </ul>
         </div>
@@ -72,19 +72,19 @@
         <div class="d-grid gap-2 m-2">
           <a
             class="btn btn-outline-primary btn-lg btn-block"
-            href="admin/currency/index.html"
+            href="admin/currency/index"
           >
             Валюты
           </a>
           <a
             class="btn btn-outline-primary btn-lg btn-block"
-            href="admin/currency_pair/index.html"
+            href="admin/currency_pair/index"
           >
             Валютные пары
           </a>
           <a
             class="btn btn-outline-primary btn-lg btn-block"
-            href="admin/exchange_rate/index.html"
+            href="admin/exchange_rate/index"
           >
             Обменный курс
           </a>
